+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.1.0
 [2022-01-30]
 * initial release.
+
+## 0.2.0
+[2022-02-05]
+### Added
+- improved channel filter performance
+- event filtering by type
+- new widget builder (MicroAppWidgetBuilder)
+- bug fixes
+- refactoring